requirements.txt
phidata
python-dotenv
yfinance
packaging
fastapi
uvicorn
groq
duckduckgo-search
openai